Version 1.2

+++ clean_all.sh
@@ -16,15 +16,23 @@ kubectl delete -f superuser-secret.yaml
 kubectl delete secret aws-creds
 kubectl delete secret minio-creds
 
-# Docker
-# Remove minio bucket
-#cat remove_minio_bucket.sh | \
-#docker run --name my-mc --hostname my-mc -e hostname=`hostname` -i --entrypoint /bin/bash --rm minio/mc
 
-docker rmi -f minio/minio
-#docker rmi -f minio/mc
+if [ "${OBJECT_STORAGE}" == "MINIO" ]; then
+ # Docker
+  # Remove minio bucket
+  #cat remove_minio_bucket.sh | \
+  #docker run --name my-mc --hostname my-mc -e hostname=`hostname` -i --entrypoint /bin/bash --rm minio/mc
+
+  docker rmi -f minio/minio
+  #docker rmi -f minio/mc
+
+  # Stop MinIO
+  docker ps | grep minio | awk '{print $1}' | xargs -I % docker stop %
+fi
 
 # AWS delete
-aws s3 rm --recursive ${s3_cluster1}
-aws s3 rm --recursive ${s3_cluster2}
+if [ "${OBJECT_STORAGE}" == "AWS" ]; then
+  aws s3 rm --recursive ${s3_cluster1}
+  aws s3 rm --recursive ${s3_cluster2}
+fi
